Move dark theme and send button into new options section
