+++ Wire-iOS/Sources/UserInterface/Settings/CellDescriptors/SettingsCellDescriptor.swift
@@ -247,7 +247,7 @@ func SettingsPropertyLabelText(_ name: SettingsPropertyName) -> String {
     case .accentColor:
         return "self.settings.account_picture_group.color".localized
     case .disableSendButton:
-        return "self.settings.send_button.title".localized
+        return "self.settings.popular_demand.send_button.title".localized
     }
 }
